# docs/swagger.yaml
definitions:
  authz-service_internal_model.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  authz-service_internal_model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/authz-service_internal_model.APIError'
    type: object
  authz-service_internal_model.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      resource_type:
        type: string
    type: object
...
  authz-service_internal_model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/v1/actions:
    get:
      description: Return all registered permission (action, resource_type) pairs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/authz-service_internal_model.Permission'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: List all registered actions
      tags:
      - admin-actions
    post:
      consumes:
      - application/json
      description: Create a new permission entry (action + resource_type). Async —
        returns 202.
      parameters:
      - description: Action definition
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Register a new action
      tags:
      - admin-actions
  /admin/v1/actions/{action}/{resource_type}:
    delete:
      description: Delete a registered permission. Async — returns 202.
      parameters:
      - description: Action name
        in: path
        name: action
        required: true
        type: string
      - description: Resource type
        in: path
        name: resource_type
        required: true
        type: string
      responses:
        "202":
          description: Accepted
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Remove an action
      tags:
      - admin-actions
  /admin/v1/roles:
    get:
      description: Return all registered roles.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/authz-service_internal_model.Role'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: List all roles
      tags:
      - admin-roles
    post:
      consumes:
      - application/json
      description: Create a new role. Async — returns 202.
      parameters:
      - description: Role definition
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Create a role
      tags:
      - admin-roles
  /admin/v1/roles/{name}/permissions:
    get:
      description: Return all permissions assigned to a role.
      parameters:
      - description: Role name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/authz-service_internal_model.Permission'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: List role permissions
      tags:
      - admin-roles
    post:
      consumes:
      - application/json
      description: Add a permission to a role. Async — returns 202.
      parameters:
      - description: Role name
        in: path
        name: name
        required: true
        type: string
      - description: Permission
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Assign permission to role
      tags:
      - admin-roles
  /admin/v1/roles/{name}/permissions/{action}/{resource_type}:
    delete:
      description: Remove a permission from a role. Async — returns 202.
      parameters:
      - description: Role name
        in: path
        name: name
        required: true
        type: string
      - description: Action name
        in: path
        name: action
        required: true
        type: string
      - description: Resource type
        in: path
        name: resource_type
        required: true
        type: string
      responses:
        "202":
          description: Accepted
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Remove permission from role
      tags:
      - admin-roles
  /admin/v1/users/{user_id}/overrides:
    post:
      consumes:
      - application/json
      description: Set an ALLOW or DENY override for a user on a specific resource.
        Async — returns 202.
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: Override definition
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Set permission override
      tags:
      - admin-overrides
  /admin/v1/users/{user_id}/overrides/{action}/{resource_type}/{resource_id}:
    delete:
      description: Remove a user's permission override for a specific resource. Async
        — returns 202.
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: Action name
        in: path
        name: action
        required: true
        type: string
      - description: Resource type
        in: path
        name: resource_type
        required: true
        type: string
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Remove permission override
      tags:
      - admin-overrides
  /admin/v1/users/{user_id}/roles:
    post:
      consumes:
      - application/json
      description: Assign a role to a user, scoped to a resource. Async — returns
        202.
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: Role assignment
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Assign role to user
      tags:
      - admin-user-roles
  /admin/v1/users/{user_id}/roles/{role_name}/{resource_type}/{resource_id}:
    delete:
      description: Revoke a user's role assignment for a specific resource. Async
        — returns 202.
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: Role name
        in: path
        name: role_name
        required: true
        type: string
      - description: Resource type
        in: path
        name: resource_type
        required: true
        type: string
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      security:
      - AdminToken: []
      summary: Revoke role from user
      tags:
      - admin-user-roles
...
swagger: "2.0"